refactor(table): move column test out of table
