+++ packages/table/__tests__/table.spec.ts
@@ -1,4 +1,4 @@
-import { mount as _mount, VueWrapper } from '@vue/test-utils'
+import { VueWrapper } from '@vue/test-utils'
 import { ComponentPublicInstance, nextTick } from 'vue'
 import ElTable from '../src/table.vue'
 import ElTableColumn from '../src/table-column/index'
@@ -529,583 +529,6 @@ describe('Table.vue', () => {
       done()
     })
   })
-  describe('column attributes', () => {
-    const createTable = function(
-      props1?,
-      props2?,
-      props3?,
-      props4?,
-      opts?,
-      tableProps?,
-    ) {
-      return mount(
-        Object.assign(
-          {
-            components: {
-              ElTable,
-              ElTableColumn,
-            },
-            template: `
-          <el-table :data="testData" ${tableProps || ''}>
-            <el-table-column prop="name" ${props1 || ''} />
-            <el-table-column prop="release" ${props2 || ''} />
-            <el-table-column prop="director" ${props3 || ''} />
-            <el-table-column prop="runtime" ${props4 || ''} />
-          </el-table>
-        `,
-
-            created() {
-              this.testData = getTestData()
-            },
-          },
-          opts,
-        ),
-      )
-    }
-
-    it('label', async done => {
-      const wrapper = createTable('label="啊哈哈哈"', 'label="啊啦啦啦"')
-      await nextTick()
-      const ths = toArray(wrapper.findAll('thead th'))
-        .map(node => node.text())
-        .filter(o => o)
-
-      expect(ths).toEqual(['啊哈哈哈', '啊啦啦啦'])
-      wrapper.unmount()
-      done()
-    })
-
-    it('width', async done => {
-      const wrapper = createTable('width="123px"', ':width="102"', 'width="39"')
-      await nextTick()
-      const ths = toArray(wrapper.findAll('.el-table__header-wrapper col'))
-        .map(node => node.attributes('width'))
-        .filter(o => o)
-      expect(ths).toContain('123')
-      expect(ths).toContain('102')
-      expect(ths).toContain('39')
-      wrapper.unmount()
-      done()
-    })
-
-    it('fixed', async done => {
-      const wrapper = createTable(
-        'fixed label="test1"',
-        'fixed="right" label="test2"',
-        'fixed="left" label="test3"',
-      )
-      await nextTick()
-      expect(
-        toArray(
-          wrapper.findAll('.el-table__fixed th:not(.is-hidden)'),
-        ).map(node => node.text()),
-      ).toEqual(['test1', 'test3'])
-
-      expect(
-        toArray(
-          wrapper.findAll('.el-table__fixed-right th:not(.is-hidden)'),
-        ).map(node => node.text()),
-      ).toEqual(['test2'])
-      expect(
-        wrapper.find('.el-table__body-wrapper').attributes('style'),
-      ).toBeFalsy()
-      wrapper.unmount()
-      done()
-    })
-
-    it('resizable', async done => {
-      const wrapper = createTable(
-        'resizable',
-        ':resizable="false"',
-        '',
-        '',
-        {},
-        'border',
-      )
-      await nextTick()
-      const firstCol = wrapper.find('thead th')
-      triggerEvent(firstCol.element, 'mousemove')
-      triggerEvent(firstCol.element, 'mousedown')
-      wrapper.unmount()
-      done()
-    })
-
-    it('formatter', async done => {
-      const wrapper = createTable(':formatter="renderCell"', '', '', '', {
-        methods: {
-          renderCell(row) {
-            return `[${row.name}]`
-          },
-        },
-      })
-
-      await nextTick()
-      const cells = toArray(
-        wrapper.findAll('.el-table__body-wrapper tbody tr td:first-child'),
-      )
-      expect(cells.map(n => n.text())).toEqual(
-        getTestData().map(o => `[${o.name}]`),
-      )
-      wrapper.unmount()
-      done()
-    })
-
-    it('show-overflow-tooltip', async done => {
-      const wrapper = createTable('show-overflow-tooltip')
-      await nextTick()
-      expect(wrapper.findAll('.el-tooltip').length).toEqual(5)
-      wrapper.unmount()
-      done()
-    })
-
-    // it('show-tooltip-when-overflow', async (done) => {
-    //   // old version prop name
-    //   const wrapper = createTable('show-tooltip-when-overflow')
-    //   await nextTick()
-    //   expect(wrapper.findAll('.el-tooltip').length).toEqual(5)
-    //   wrapper.unmount()
-    //   done()
-    // })
-
-    it('align', async done => {
-      const wrapper = createTable(
-        'align="left"',
-        'align="right"',
-        'align="center"',
-      )
-      await nextTick()
-      const len = getTestData().length + 1
-      expect(wrapper.findAll('.is-left').length).toEqual(len)
-      expect(wrapper.findAll('.is-right').length).toEqual(len)
-      expect(wrapper.findAll('.is-center').length).toEqual(len)
-      wrapper.unmount()
-      done()
-    })
-
-    it('class-name', async done => {
-      const wrapper = createTable(
-        'class-name="column-1"',
-        'class-name="column-2 column-class-a"',
-        'class-name="column-class-a"',
-      )
-      await nextTick()
-      const len = getTestData().length + 1
-      expect(wrapper.findAll('.column-1').length).toEqual(len)
-      expect(wrapper.findAll('.column-2').length).toEqual(len)
-      expect(wrapper.findAll('.column-class-a').length).toEqual(len * 2)
-      wrapper.unmount()
-      done()
-    })
-
-    it('selectable === false & check selectAll status', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData" @selection-change="change">
-            <el-table-column type="selection" :selectable="filterSelect" />
-            <el-table-column prop="name" label="name" />
-            <el-table-column prop="release" label="release" />
-            <el-table-column prop="director" label="director" />
-            <el-table-column prop="runtime" label="runtime" />
-          </el-table>
-        `,
-
-        data() {
-          return { selected: [], testData: getTestData() }
-        },
-
-        methods: {
-          change(rows) {
-            this.selected = rows
-          },
-
-          filterSelect() {
-            return false
-          },
-        },
-      })
-
-      await nextTick()
-      expect(wrapper.find('.el-checkbox').attributes('checked')).toBeFalsy()
-      await nextTick()
-      expect(wrapper.vm.selected.length).toEqual(0)
-      wrapper.unmount()
-      done()
-    })
-
-    describe('type', () => {
-      const createTable = function(type) {
-        return mount({
-          components: {
-            ElTable,
-            ElTableColumn,
-          },
-          template: `
-            <el-table :data="testData" @selection-change="change">
-              <el-table-column type="${type}" />
-              <el-table-column prop="name" label="name" />
-              <el-table-column prop="release" label="release" />
-              <el-table-column prop="director" label="director" />
-              <el-table-column prop="runtime" label="runtime" />
-            </el-table>
-          `,
-
-          created() {
-            this.testData = getTestData()
-          },
-
-          data() {
-            return { selected: [] }
-          },
-
-          methods: {
-            change(rows) {
-              this.selected = rows
-            },
-          },
-        })
-      }
-
-      describe('= selection', () => {
-        const wrapper = createTable('selection')
-
-        it('render', async done => {
-          await nextTick()
-          expect(wrapper.findAll('.el-checkbox').length).toEqual(
-            getTestData().length + 1,
-          )
-          done()
-        })
-
-        it('select all', async done => {
-          wrapper.find('.el-checkbox').trigger('click')
-          await nextTick()
-          expect(wrapper.vm.selected.length).toEqual(5)
-          wrapper.unmount()
-          done()
-        })
-
-        it('select one', async done => {
-          const wrapper2 = createTable('selection')
-
-          await nextTick()
-          wrapper2.findAll('.el-checkbox')[1].trigger('click')
-
-          await nextTick()
-          expect(wrapper2.vm.selected.length).toEqual(1)
-          expect(wrapper2.vm.selected[0].name).toEqual(getTestData()[0].name)
-          wrapper2.unmount()
-          done()
-        })
-      })
-
-      describe('= index', () => {
-        const wrapper = createTable('index')
-
-        it('render', async done => {
-          await nextTick()
-          expect(
-            toArray(
-              wrapper.findAll(
-                '.el-table__body-wrapper tbody tr td:first-child',
-              ),
-            ).map(node => node.text()),
-          ).toEqual(['1', '2', '3', '4', '5'])
-          wrapper.unmount()
-          done()
-        })
-      })
-
-      describe('= expand', () => {
-        const createInstance = function(extra?) {
-          extra = extra || ''
-          return mount({
-            components: {
-              ElTableColumn,
-              ElTable,
-            },
-            template: `
-            <el-table row-key="id" :data="testData" @expand-change="handleExpand" ${extra}>
-              <el-table-column type="expand">
-                <template #default="props">
-                  <div>{{props.row.name}}</div>
-                </template>
-              </el-table-column>
-              <el-table-column prop="release" label="release" />
-              <el-table-column prop="director" label="director" />
-              <el-table-column prop="runtime" label="runtime" />
-            </el-table>
-          `,
-
-            data() {
-              return {
-                expandCount: 0,
-                expandRowKeys: [],
-                testData: getTestData(),
-              }
-            },
-
-            methods: {
-              handleExpand() {
-                this.expandCount++
-              },
-              refreshData() {
-                this.testData = getTestData()
-              },
-            },
-          })
-        }
-
-        it('works', async done => {
-          const wrapper = createInstance()
-          await nextTick()
-          expect(wrapper.findAll('td.el-table__expand-column').length).toEqual(
-            5,
-          )
-          wrapper.unmount()
-          done()
-        })
-      })
-    })
-
-    describe('sortable', () => {
-      it('render', async done => {
-        const wrapper = createTable('', '', '', 'sortable')
-        await nextTick()
-        expect(wrapper.findAll('.caret-wrapper').length).toEqual(1)
-        wrapper.unmount()
-        done()
-      })
-
-      it('sortable orders', async done => {
-        const wrapper = createTable(
-          '',
-          '',
-          '',
-          "sortable :sort-orders=\"['descending', 'ascending']\"",
-          {},
-        )
-
-        await nextTick()
-        const elm = wrapper.find('.caret-wrapper')
-        elm.trigger('click')
-
-        await nextTick()
-        const lastCells = wrapper.findAll(
-          '.el-table__body-wrapper tbody tr td:last-child',
-        )
-        expect(toArray(lastCells).map(node => node.text())).toEqual([
-          '100',
-          '95',
-          '92',
-          '92',
-          '80',
-        ])
-        wrapper.unmount()
-        done()
-      })
-
-      it('sortable method', async done => {
-        const wrapper = createTable(
-          'sortable :sort-method="sortMethod"',
-          '',
-          '',
-          '',
-          {
-            methods: {
-              sortMethod(a, b) {
-                // sort method should return number
-                if (a.runtime < b.runtime) {
-                  return 1
-                }
-                if (a.runtime > b.runtime) {
-                  return -1
-                }
-                return 0
-              },
-            },
-          },
-        )
-
-        await nextTick()
-        const elm = wrapper.find('.caret-wrapper')
-        elm.trigger('click')
-
-        await nextTick()
-        const lastCells = wrapper.findAll(
-          '.el-table__body-wrapper tbody tr td:last-child',
-        )
-        expect(toArray(lastCells).map(node => node.text())).toEqual([
-          '100',
-          '95',
-          '92',
-          '92',
-          '80',
-        ])
-        wrapper.unmount()
-        done()
-      })
-
-      it('sortable by method', async done => {
-        const wrapper = createTable('sortable :sort-by="sortBy"', '', '', '', {
-          methods: {
-            sortBy(a) {
-              return -a.runtime
-            },
-          },
-        })
-
-        await nextTick()
-        const elm = wrapper.find('.caret-wrapper')
-        elm.trigger('click')
-
-        await nextTick()
-        const lastCells = wrapper.findAll(
-          '.el-table__body-wrapper tbody tr td:last-child',
-        )
-        expect(toArray(lastCells).map(node => node.text())).toEqual([
-          '100',
-          '95',
-          '92',
-          '92',
-          '80',
-        ])
-        wrapper.unmount()
-        done()
-      })
-
-      it('sortable by property', async done => {
-        const wrapper = createTable(
-          'sortable sort-by="runtime"',
-          '',
-          '',
-          '',
-          {},
-        )
-
-        await nextTick()
-        const elm = wrapper.find('.caret-wrapper')
-        elm.trigger('click')
-
-        await nextTick()
-        const lastCells = wrapper.findAll(
-          '.el-table__body-wrapper tbody tr td:last-child',
-        )
-        expect(toArray(lastCells).map(node => node.text())).toEqual([
-          '80',
-          '92',
-          '92',
-          '95',
-          '100',
-        ])
-        wrapper.unmount()
-        done()
-      })
-    })
-
-    describe('click sortable column', () => {
-      const wrapper = createTable('', '', '', 'sortable')
-
-      it('ascending', async done => {
-        const elm = wrapper.find('.caret-wrapper')
-
-        elm.trigger('click')
-        await nextTick()
-        const lastCells = wrapper.findAll(
-          '.el-table__body-wrapper tbody tr td:last-child',
-        )
-        expect(toArray(lastCells).map(node => node.text())).toEqual([
-          '80',
-          '92',
-          '92',
-          '95',
-          '100',
-        ])
-        done()
-      })
-
-      it('descending', async done => {
-        const elm = wrapper.find('.caret-wrapper')
-
-        elm.trigger('click')
-        await nextTick()
-        const lastCells = wrapper.findAll(
-          '.el-table__body-wrapper tbody tr td:last-child',
-        )
-        expect(toArray(lastCells).map(node => node.text())).toEqual([
-          '100',
-          '95',
-          '92',
-          '92',
-          '80',
-        ])
-        wrapper.unmount()
-        done()
-      })
-    })
-
-    it('change column configuration', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <template>
-            <button
-              id="addBut"
-              @click="
-                () => {
-                  cols.push('b')
-                }
-              "
-              >+</button>
-            <button
-              id="delBut"
-              @click="
-                () => {
-                  cols.pop()
-                }
-              "
-              >-</button>
-            <el-table :data="data">
-              <el-table-column
-                v-for="item of cols"
-                :prop="item"
-                :label="item"
-                :key="item"
-              ></el-table-column>
-            </el-table>
-          </template>
-        `,
-
-        data() {
-          return { cols: ['a', 'v', 'b'], data: [{ a: 1, v: 2, b: 3 }] }
-        },
-      })
-
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header-wrapper th').length).toEqual(3)
-      const addBut = wrapper.find('#addBut')
-      const delBut = wrapper.find('#delBut')
-      addBut.trigger('click')
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header-wrapper th').length).toEqual(4)
-      addBut.trigger('click')
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header-wrapper th').length).toEqual(5)
-      delBut.trigger('click')
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header-wrapper th').length).toEqual(4)
-      delBut.trigger('click')
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header-wrapper th').length).toEqual(3)
-      done()
-    })
-  })
   describe('summary row', () => {
     it('should render', async done => {
       const wrapper = mount({
@@ -1208,404 +631,6 @@ describe('Table.vue', () => {
       await nextTick()
       const cells = toArray(wrapper.findAll('.el-table__footer .cell'))
       expect(cells[1].text()).toEqual('9996')
-      wrapper.unmount()
-      done()
-    })
-  })
-  describe('multi level column', () => {
-    it('should works', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column prop="name" />
-            <el-table-column label="group">
-              <el-table-column prop="release"/>
-              <el-table-column prop="director"/>
-            </el-table-column>
-            <el-table-column prop="runtime"/>
-          </el-table>
-        `,
-
-        created() {
-          this.testData = null
-        },
-      })
-
-      await nextTick()
-      const trs = wrapper.findAll('.el-table__header tr')
-      expect(trs.length).toEqual(2)
-      const firstRowHeader = trs[0].findAll('th .cell').length
-      const secondRowHeader = trs[1].findAll('th .cell').length
-      expect(firstRowHeader).toEqual(3)
-      expect(secondRowHeader).toEqual(2)
-
-      expect(trs[0].find('th:first-child').attributes('rowspan')).toEqual('2')
-      expect(trs[0].find('th:nth-child(2)').attributes('colspan')).toEqual('2')
-      wrapper.unmount()
-      done()
-    })
-    // TODO
-    // it('should works', async done => {
-    //   const wrapper = mount({
-    //     components: {
-    //       ElTable,
-    //       ElTableColumn,
-    //     },
-    //     template: `
-    //       <el-table :data="testData">
-    //         <el-table-column prop="name" />
-    //         <el-table-column label="group">
-    //           <el-table-column label="group's group">
-    //             <el-table-column prop="release" />
-    //             <el-table-column prop="runtime"/>
-    //           </el-table-column>
-    //           <el-table-column prop="director" />
-    //         </el-table-column>
-    //         <el-table-column prop="runtime"/>
-    //       </el-table>
-    //     `,
-
-    //     created() {
-    //       this.testData = null
-    //     },
-    //   })
-
-    //   await nextTick()
-    //   const trs = wrapper.findAll('.el-table__header tr')
-    //   expect(trs.length).toEqual(3)
-    //   const firstRowHeader = trs[0].findAll('th .cell').length
-    //   const secondRowHeader = trs[1].findAll('th .cell').length
-    //   const thirdRowHeader = trs[2].findAll('th .cell').length
-    //   expect(firstRowHeader).toEqual(3)
-    //   expect(secondRowHeader).toEqual(2)
-    //   expect(thirdRowHeader).toEqual(2)
-
-    //   expect(trs[0].find('th:first-child').attributes('rowspan')).toEqual('3')
-    //   expect(trs[0].find('th:nth-child(2)').attributes('colspan')).toEqual('3')
-    //   expect(trs[1].find('th:first-child').attributes('colspan')).toEqual('2')
-    //   expect(trs[1].find('th:nth-child(2)').attributes('rowspan')).toEqual('2')
-
-    //   wrapper.unmount()
-    //   done()
-    // })
-
-    it('should work in one column', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column label="group">
-              <el-table-column prop="release"/>
-            </el-table-column>
-          </el-table>
-        `,
-
-        created() {
-          this.testData = null
-        },
-      })
-
-      await nextTick()
-      const trs = wrapper.findAll('.el-table__header tr')
-      expect(trs.length).toEqual(2)
-      const firstRowLength = trs[0].findAll('th .cell').length
-      const secondRowLength = trs[1].findAll('th .cell').length
-      expect(firstRowLength).toEqual(1)
-      expect(secondRowLength).toEqual(1)
-
-      expect(trs[0].find('th:first-child').attributes('rowspan')).toEqual('1')
-      expect(trs[0].find('th:first-child').attributes('colspan')).toEqual('1')
-      wrapper.unmount()
-      done()
-    })
-  })
-  describe('dynamic column attribtes', () => {
-    it('label', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column prop="name" :label="label"/>
-            <el-table-column prop="release" />
-            <el-table-column prop="director" />
-            <el-table-column prop="runtime" />
-          </el-table>
-        `,
-        data() {
-          return {
-            label: 'name',
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-
-      await nextTick()
-      expect(wrapper.find('.el-table__header th .cell').text()).toEqual('name')
-      wrapper.vm.label = 'NAME'
-      wrapper.vm.$nextTick(() => {
-        expect(wrapper.find('.el-table__header th .cell').text()).toEqual(
-          'NAME',
-        )
-        wrapper.unmount()
-        done()
-      })
-    })
-
-    it('align', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column prop="name" :align="align"/>
-          </el-table>
-        `,
-
-        data() {
-          return {
-            align: 'left',
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-
-      await nextTick()
-      expect(wrapper.findAll('.el-table__body td.is-right').length).toEqual(0)
-      wrapper.vm.align = 'right'
-      wrapper.vm.$nextTick(() => {
-        expect(
-          wrapper.findAll('.el-table__body td.is-right').length > 0,
-        ).toBeTruthy()
-        wrapper.unmount()
-        done()
-      })
-    })
-    it('header-align', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column prop="name" :align="align" :header-align="headerAlign"/>
-          </el-table>
-        `,
-
-        data() {
-          return {
-            align: 'left',
-            headerAlign: null,
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-      await nextTick()
-      expect(
-        wrapper.findAll('.el-table__header th.is-left').length,
-      ).toBeGreaterThanOrEqual(0)
-      expect(wrapper.findAll('.el-table__header th.is-center').length).toEqual(
-        0,
-      )
-      expect(wrapper.findAll('.el-table__header th.is-right').length).toEqual(0)
-      wrapper.vm.align = 'right'
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header th.is-left').length).toEqual(0)
-      expect(wrapper.findAll('.el-table__header th.is-center').length).toEqual(
-        0,
-      )
-      expect(
-        wrapper.findAll('.el-table__header th.is-right').length,
-      ).toBeGreaterThanOrEqual(0)
-      wrapper.vm.headerAlign = 'center'
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header th.is-left').length).toEqual(0)
-      expect(
-        wrapper.findAll('.el-table__header th.is-center').length,
-      ).toBeGreaterThanOrEqual(0)
-      expect(wrapper.findAll('.el-table__header th.is-right').length).toEqual(0)
-      wrapper.vm.headerAlign = null
-      await nextTick()
-      expect(wrapper.findAll('.el-table__header th.is-left').length).toEqual(0)
-      expect(wrapper.findAll('.el-table__header th.is-center').length).toEqual(
-        0,
-      )
-      expect(
-        wrapper.findAll('.el-table__header th.is-right').length,
-      ).toBeGreaterThanOrEqual(0)
-      wrapper.unmount()
-      done()
-    })
-
-    it('width', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData" :fit="false">
-            <el-table-column prop="name" :width="width"/>
-          </el-table>
-        `,
-
-        data() {
-          return {
-            width: 100,
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-
-      await nextTick()
-      expect(wrapper.find('.el-table__body col').attributes('width')).toEqual(
-        '100',
-      )
-      wrapper.vm.width = 200
-      await nextTick()
-      expect(wrapper.find('.el-table__body col').attributes('width')).toEqual(
-        '200',
-      )
-      wrapper.unmount()
-      done()
-    })
-
-    it('min-width', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData" :fit="false">
-            <el-table-column prop="name" :min-width="width"/>
-          </el-table>
-        `,
-
-        data() {
-          return {
-            width: 100,
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-
-      await nextTick()
-      expect(wrapper.find('.el-table__body col').attributes('width')).toEqual(
-        '100',
-      )
-      wrapper.vm.width = 200
-      await nextTick()
-      expect(wrapper.find('.el-table__body col').attributes('width')).toEqual(
-        '200',
-      )
-      wrapper.unmount()
-      done()
-    })
-
-    it('fixed', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column :fixed="fixed" />
-            <el-table-column prop="release" />
-            <el-table-column prop="director" />
-            <el-table-column prop="runtime" />
-          </el-table>
-        `,
-
-        data() {
-          return {
-            fixed: false,
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-
-      await nextTick()
-      expect(Object.keys(wrapper.find('.el-table__fixed')).length).toEqual(0)
-      wrapper.vm.fixed = true
-      await nextTick()
-      expect(
-        Object.keys(wrapper.find('.el-table__fixed')).length,
-      ).toBeGreaterThan(0)
-      wrapper.unmount()
-      done()
-    })
-
-    it('prop', async done => {
-      const wrapper = mount({
-        components: {
-          ElTable,
-          ElTableColumn,
-        },
-        template: `
-          <el-table :data="testData">
-            <el-table-column :prop="prop" />
-            <el-table-column prop="release" />
-            <el-table-column prop="director" />
-            <el-table-column prop="runtime" />
-          </el-table>
-        `,
-
-        data() {
-          return {
-            prop: 'name',
-          }
-        },
-
-        created() {
-          this.testData = getTestData()
-        },
-      })
-
-      await nextTick()
-      let firstColumnContent = wrapper.find('.el-table__body td .cell').text()
-      let secondColumnContent = wrapper
-        .find('.el-table__body td:nth-child(2) .cell')
-        .text()
-      expect(firstColumnContent).not.toEqual(secondColumnContent)
-      wrapper.vm.prop = 'release'
-      await nextTick()
-      firstColumnContent = wrapper.find('.el-table__body td .cell').text()
-      secondColumnContent = wrapper
-        .find('.el-table__body td:nth-child(2) .cell')
-        .text()
-      expect(firstColumnContent).toEqual(secondColumnContent)
       wrapper.unmount()
       done()
     })
